Move further structs into correct headers
